+++ src/templates/rainier.js
@@ -1,12 +1,12 @@
 import React from 'react'
-import './rainer-styles.css'
+import './rainier-styles.css'
 import Warning from '../warning'
 
 const ALLOWED_FORMATS = ['inline', 'modal', 'slide in']
 
 const bgImage = '//pages.convertkit.com/assets/rainier/bg.jpg'
 
-function Rainer({
+function Rainier({
   options,
   className,
   action,
@@ -26,7 +26,7 @@ function Rainer({
   return (
     <>
       {!hideWarnings && !ALLOWED_FORMATS.includes(format) ? <Warning message="This template is not available for the chosen format" /> : null}
-      <form action={action} className={`seva-form formkit-form rainer ${className}`} method="post" target={newTab ? '_blank' : '_self'} data-sv-form={formId} data-uid={options.uid} data-format={format} data-version={options.version} data-options={JSON.stringify(options)} min-width="400 500 600 700 800">
+      <form action={action} className={`seva-form formkit-form rainier ${className}`} method="post" target={newTab ? '_blank' : '_self'} data-sv-form={formId} data-uid={options.uid} data-format={format} data-version={options.version} data-options={JSON.stringify(options)} min-width="400 500 600 700 800">
         <div className="formkit-container" data-style="full" style={{ backgroundColor: 'rgb(255, 255, 255)' }}>
           <div className="formkit-hero">
             <div className="formkit-hero-bg-color" style={{ background: `linear-gradient(rgba(16, 16, 16, .8), rgba(16, 16, 16, .8)), url(${bgImage})`, backgroundSize: 'cover', backgroundPosition: 'center' }}></div>
@@ -72,4 +72,4 @@ function Rainer({
   )
 }
 
-export default Rainer
+export default Rainier